chore: move utility types out
