+++ src/lib/swapper/swappers/CowSwapper/getCowSwapTradeQuote/getCowSwapTradeQuote.ts
@@ -1,5 +1,6 @@
 import { fromAssetId } from '@shapeshiftoss/caip'
 import type { SwapErrorRight } from '@shapeshiftoss/swapper'
+import { createTradeAmountTooSmallErr } from '@shapeshiftoss/swapper'
 import type { Result } from '@sniptt/monads'
 import { Err, Ok } from '@sniptt/monads'
 import type { AxiosError } from 'axios'
@@ -26,7 +27,6 @@ import {
 } from 'lib/swapper/swappers/utils/helpers/helpers'
 import type { GetTradeQuoteInput, TradeQuote } from 'lib/swapper/types'
 import { SwapperName } from 'lib/swapper/types'
-import { createTradeAmountTooSmallErr } from 'lib/swapper/utils'
 
 export async function getCowSwapTradeQuote(
   input: GetTradeQuoteInput,
