+++ src/lib/swapper/swappers/CowSwapper/utils/helpers/helpers.ts
@@ -4,7 +4,7 @@ import type { OrderClass } from '@cowprotocol/app-data/dist/generatedTypes/v0.9.
 import type { TypedDataDomain, TypedDataField } from '@ethersproject/abstract-signer'
 import type { ChainId } from '@shapeshiftoss/caip'
 import { fromAssetId } from '@shapeshiftoss/caip'
-import { type SwapErrorRight, SwapErrorType } from '@shapeshiftoss/swapper'
+import { makeSwapErrorRight, type SwapErrorRight, SwapErrorType } from '@shapeshiftoss/swapper'
 import { KnownChainIds } from '@shapeshiftoss/types'
 import type { Result } from '@sniptt/monads'
 import { Err, Ok } from '@sniptt/monads'
@@ -15,7 +15,6 @@ import type { Asset } from 'lib/asset-service'
 import { bnOrZero, convertPrecision } from 'lib/bignumber/bignumber'
 import { fromBaseUnit } from 'lib/math'
 import type { CowSwapOrder } from 'lib/swapper/types'
-import { makeSwapErrorRight } from 'lib/swapper/utils'
 
 import type { CowSwapQuoteResponse } from '../../types'
 import { CowNetwork } from '../../types'
