+++ src/lib/swapper/swappers/LifiSwapper/getTradeQuote/getTradeQuote.ts
@@ -2,7 +2,7 @@ import type { ChainKey, LifiError, RoutesRequest } from '@lifi/sdk'
 import { LifiErrorCode } from '@lifi/sdk'
 import type { AssetId, ChainId } from '@shapeshiftoss/caip'
 import { fromChainId } from '@shapeshiftoss/caip'
-import { type SwapErrorRight, SwapErrorType } from '@shapeshiftoss/swapper'
+import { makeSwapErrorRight, type SwapErrorRight, SwapErrorType } from '@shapeshiftoss/swapper'
 import type { Result } from '@sniptt/monads'
 import { Err, Ok } from '@sniptt/monads'
 import { getConfig } from 'config'
@@ -17,7 +17,6 @@ import { transformLifiStepFeeData } from 'lib/swapper/swappers/LifiSwapper/utils
 import type { LifiTradeQuote } from 'lib/swapper/swappers/LifiSwapper/utils/types'
 import type { GetEvmTradeQuoteInput, SwapSource } from 'lib/swapper/types'
 import { SwapperName } from 'lib/swapper/types'
-import { makeSwapErrorRight } from 'lib/swapper/utils'
 import { isFulfilled } from 'lib/utils'
 import { convertBasisPointsToDecimalPercentage } from 'state/slices/tradeQuoteSlice/utils'
