+++ src/lib/swapper/swappers/ThorchainSwapper/cosmossdk/getCosmosTxData.ts
@@ -3,7 +3,7 @@ import { cosmosAssetId } from '@shapeshiftoss/caip'
 import type { CosmosSdkBaseAdapter, thorchain } from '@shapeshiftoss/chain-adapters'
 import type { CosmosSignTx, ThorchainSignTx } from '@shapeshiftoss/hdwallet-core'
 import type { SwapErrorRight } from '@shapeshiftoss/swapper'
-import { SwapErrorType } from '@shapeshiftoss/swapper'
+import { makeSwapErrorRight, SwapErrorType } from '@shapeshiftoss/swapper'
 import { KnownChainIds } from '@shapeshiftoss/types'
 import type { Result } from '@sniptt/monads'
 import { Err, Ok } from '@sniptt/monads'
@@ -11,7 +11,6 @@ import { getConfig } from 'config'
 import type { Asset } from 'lib/asset-service'
 import type { ThorCosmosSdkSupportedChainId } from 'lib/swapper/swappers/ThorchainSwapper/types'
 import type { CosmosSdkFeeData, TradeQuote } from 'lib/swapper/types'
-import { makeSwapErrorRight } from 'lib/swapper/utils'
 import { getInboundAddressDataForChain } from 'lib/utils/thorchain/getInboundAddressDataForChain'
 
 type GetCosmosTxDataInput = {
