+++ src/lib/swapper/swappers/ThorchainSwapper/utils/getL1quote.ts
@@ -1,7 +1,12 @@
 import type { AssetId } from '@shapeshiftoss/caip'
 import { CHAIN_NAMESPACE, fromAssetId } from '@shapeshiftoss/caip'
 import { bn, bnOrZero } from '@shapeshiftoss/chain-adapters'
-import { type SwapErrorRight, SwapErrorType } from '@shapeshiftoss/swapper'
+import {
+  createTradeAmountTooSmallErr,
+  makeSwapErrorRight,
+  type SwapErrorRight,
+  SwapErrorType,
+} from '@shapeshiftoss/swapper'
 import { Err, Ok, type Result } from '@sniptt/monads'
 import { getConfig } from 'config'
 import { getDefaultSlippageDecimalPercentageForSwapper } from 'constants/constants'
@@ -12,7 +17,6 @@ import { getThorTxInfo as getEvmThorTxInfo } from 'lib/swapper/swappers/Thorchai
 import { getThorTxInfo as getUtxoThorTxInfo } from 'lib/swapper/swappers/ThorchainSwapper/utxo/utils/getThorTxData'
 import type { GetEvmTradeQuoteInput, GetUtxoTradeQuoteInput, ProtocolFee } from 'lib/swapper/types'
 import { type GetTradeQuoteInput, SwapperName } from 'lib/swapper/types'
-import { createTradeAmountTooSmallErr, makeSwapErrorRight } from 'lib/swapper/utils'
 import { assertUnreachable, isFulfilled, isRejected } from 'lib/utils'
 import { assertGetCosmosSdkChainAdapter } from 'lib/utils/cosmosSdk'
 import { assertGetEvmChainAdapter } from 'lib/utils/evm'
