+++ src/lib/swapper/swappers/ThorchainSwapper/utils/getQuote/getQuote.ts
@@ -1,6 +1,11 @@
 import type { AssetId } from '@shapeshiftoss/caip'
 import { bchAssetId } from '@shapeshiftoss/caip'
-import { type SwapErrorRight, SwapErrorType } from '@shapeshiftoss/swapper'
+import {
+  createTradeAmountTooSmallErr,
+  makeSwapErrorRight,
+  type SwapErrorRight,
+  SwapErrorType,
+} from '@shapeshiftoss/swapper'
 import type { Result } from '@sniptt/monads'
 import { Err, Ok } from '@sniptt/monads'
 import { getConfig } from 'config'
@@ -17,7 +22,6 @@ import {
   THORCHAIN_FIXED_PRECISION,
 } from 'lib/swapper/swappers/ThorchainSwapper/utils/constants'
 import { assetIdToPoolAssetId } from 'lib/swapper/swappers/ThorchainSwapper/utils/poolAssetHelpers/poolAssetHelpers'
-import { createTradeAmountTooSmallErr, makeSwapErrorRight } from 'lib/swapper/utils'
 
 import { getThresholdedAffiliateBps } from '../getThresholdedAffiliateBps/getThresholdedAffiliateBps'
 import { thorService } from '../thorService'
