+++ src/lib/swapper/utils.ts
@@ -1,6 +1,5 @@
-import type { AssetId } from '@shapeshiftoss/caip'
 import type { SwapErrorRight } from '@shapeshiftoss/swapper'
-import { SwapErrorType } from '@shapeshiftoss/swapper'
+import { makeSwapErrorRight, SwapErrorType } from '@shapeshiftoss/swapper'
 import type { Result } from '@sniptt/monads'
 import { Err, Ok } from '@sniptt/monads'
 import type { AxiosInstance, AxiosRequestConfig, AxiosResponse } from 'axios'
@@ -11,24 +10,6 @@ import type { FeatureFlags } from 'state/slices/preferencesSlice/preferencesSlic
 import { isCrossAccountTradeSupported } from '../../state/helpers'
 import { AsyncResultOf, isTruthy } from '../utils'
 import { SwapperName } from './types'
-
-export const makeSwapErrorRight = ({
-  details,
-  cause,
-  code,
-  message,
-}: {
-  message: string
-  details?: unknown
-  cause?: unknown
-  code?: SwapErrorType
-}): SwapErrorRight => ({
-  name: 'SwapError',
-  message,
-  details,
-  cause,
-  code,
-})
 
 const getRequestFilter = (cachedUrls: string[]) => (request: Request) =>
   !cachedUrls.some(url => request.url.includes(url))
@@ -124,13 +105,3 @@ export const getEnabledSwappers = (
       return !isCrossAccountTrade || swapperSupportsCrossAccountTrade
     })
 }
-
-export const createTradeAmountTooSmallErr = (details?: {
-  minAmountCryptoBaseUnit: string
-  assetId: AssetId
-}) =>
-  makeSwapErrorRight({
-    code: SwapErrorType.TRADE_QUOTE_AMOUNT_TOO_SMALL,
-    message: 'Sell amount is too small',
-    details,
-  })
